+++ rust/windows-client/src/settings.html
@@ -5,6 +5,7 @@
     <link rel="stylesheet" href="output.css" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Settings</title>
+    <script src="./flowbite.min.js" defer></script>
     <script type="module" src="settings.js" defer></script>
   </head>
 
@@ -176,6 +177,5 @@
         </div>
       </div>
     </div>
-    <script src="./flowbite.min.js"></script>
   </body>
 </html>
